Added Checkpoint System
To add a checkpoint, drag in a checkpoint prefab to the scene, and duplicate it to make more. Make sure each checkpoint has the number when it would occur in the scene (1 would occur 1st, 2 would occur 2nd, etc). Drag all these checkpoints onto the Player's Checkpoint manager script, in order, so they work properly). You can look in the MohsenTestingScene Scene for an example.