Refactored Player Movement
An additional "Falling" state was added to the state machine which handles the aftermath of performing a jump and walking off a ledge.
IsGrounded() has been refactored to account for the downward diagonals.
The player's jump now has a maximum airtime and will start falling either once that airtime is reached or once the player releases the jump button.